refactor: ephemeral repository
