+++ src/test/java/com/example/roommate/tests/controller/booking/PostBookTest.java
@@ -5,8 +5,8 @@ import com.example.roommate.annotations.TestClass;
 import com.example.roommate.annotations.WithMockOAuth2User;
 import com.example.roommate.domain.services.RoomDomainService;
 import com.example.roommate.exceptions.applicationService.NotFoundException;
-import com.example.roommate.persistence.data.RoomEntry;
-import com.example.roommate.persistence.repositories.RoomRepository;
+import com.example.roommate.persistence.ephemeral.RoomEntry;
+import com.example.roommate.persistence.ephemeral.RoomRepository;
 import com.example.roommate.values.forms.BookDataForm;
 import com.example.roommate.application.services.BookingApplicationService;
 import org.junit.jupiter.api.Disabled;
