+++ src/test/java/com/example/roommate/tests/services/RoomDomainServiceTest.java
@@ -3,11 +3,11 @@ package com.example.roommate.tests.services;
 import com.example.roommate.annotations.TestClass;
 import com.example.roommate.application.data.RoomApplicationData;
 import com.example.roommate.interfaces.entities.IRoom;
-import com.example.roommate.persistence.data.RoomEntry;
+import com.example.roommate.persistence.ephemeral.RoomEntry;
 import com.example.roommate.domain.models.entities.Room;
 import com.example.roommate.values.domainValues.ItemName;
-import com.example.roommate.persistence.repositories.ItemRepository;
-import com.example.roommate.persistence.repositories.RoomRepository;
+import com.example.roommate.persistence.ephemeral.ItemRepository;
+import com.example.roommate.persistence.ephemeral.RoomRepository;
 import com.example.roommate.exceptions.NotFoundRepositoryException;
 import com.example.roommate.domain.services.RoomDomainService;
 import com.example.roommate.factories.EntityFactory;
@@ -18,7 +18,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.*;
